Change call time from long to double

+++ src/main/java/com/zeroturnaround/callspy/Timer.java
@@ -3,11 +3,11 @@ import java.util.Stack;
 
 @SuppressWarnings("unused")
 public class Timer {
-    private static Stack<Long> stack = new Stack<>();
+    private static Stack<Double> stack = new Stack<>();
     static int indent = 0;
 
     public static void start() {
-        stack.push(System.nanoTime());
+        stack.push((double) System.nanoTime());
         indent++;
         System.out.print(String.format("%" + indent + "s >", ""));
     }
@@ -18,6 +18,6 @@ public class Timer {
     }
 
     public static void log() {
-        System.out.println(String.format(" - took %d ms", ((System.nanoTime() - stack.pop()) / 1_000_000)));
+        System.out.println(String.format(" - took %.2f ms", ((System.nanoTime() - stack.pop()) / 1_000_000)));
     }
 }
